apps/backend/dev-requirements.txt
debugpy==1.8.1
watchdog==4.0.0
ipdb==0.13.13
ipython==8.22.2
pytest==8.1.1
pytest-cov==4.1.0
black==24.3.0
flake8==7.0.0
aniso8601==9.0.1
asgiref==3.8.1
cffi==1.16.0
cryptography==44.0.2
Django==5.0.7
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.2
django-grappelli==4.0.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-phonenumber-field==8.0.0
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==22.0.0
packaging==24.1
phonenumbers==8.13.40
pillow==10.4.0
promise==2.3
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tzdata==2024.1
whitenoise==6.7.0
graphene-django-optimizer
django-robots
django-meta
django-resized
django-check-seo==1.0.1
django-amp-tools
